Remove condition for deleting SSM parameter for certificate
The certificate can be in different states depending on what triggered the deletion.
Making sure we delete it only when the parameter value matches the current certificate ARN is enough to guarantee we're deleting
the parameter in the right situation. This check is done inside the _delete_ssm_parameter function.